refactor(core): split temporal guide into basic axis, types, and candidates
Extract planBasicAxis into basic-axis.ts and temporal interval scoring into
temporal-axis-candidates.ts; keep temporal-guide.ts as the stable facade
with guide-plan type re-exports. Split characterization into
tests/temporal-guide/ label-formatting and measured-axis-plan suites.
