+++ packages/core/tests/temporal-guide/fixtures.ts
@@ -1,0 +1,4 @@
+import { FONT_METRICS } from "../../src/layout/font-metrics.ts";
+import { MetricsTableMeasurer } from "../../src/layout/measure.ts";
+
+export const measurer = new MetricsTableMeasurer(FONT_METRICS);
